fix: 미션 작성 내용 placeholder 수정
