+++ package.json
@@ -38,6 +38,7 @@
     "js-cookie": "^3.0.5",
     "jwt-decode": "^4.0.0",
     "next": "15.1.7",
+    "next-pwa": "^5.6.0",
     "next-seo": "^6.6.0",
     "next-themes": "^0.4.4",
     "react": "^19.0.0",
@@ -50,6 +51,7 @@
     "swr": "^2.3.3",
     "tippy.js": "^6.3.7",
     "ua-parser-js": "^2.0.3",
+    "web-push": "^3.6.7",
     "zod": "^3.24.2",
     "zustand": "^5.0.3"
   },
