+++ src/app/mission/create/client.tsx
@@ -183,5 +183,3 @@ const NewMissionPageContainer = styled.div`
     width: 100%;
   }
 `;
-
-//TODO: 1. 미션 생성 페이지 및 수정 페이지
