+++ src/components/common/Checkbox.module.css
@@ -11,8 +11,8 @@
 }
 
 .checkbox:checked {
-    background-color: #DD6B20;  /* 주황색 직접 지정 */
-    border: 1px solid #DD6B20;
+    background-color: var(--primary-500);
+    border: 1px solid var(--primary-500);
     background-image: url("data:image/svg+xml,%3Csvg width='12' height='9' viewBox='0 0 12 9' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M10.6667 1.5L4.25004 7.91667L1.33337 5' stroke='white' stroke-width='2' stroke-linecap='round' stroke-linejoin='round'/%3E%3C/svg%3E");
     background-repeat: no-repeat;
     background-position: center;
