+++ src/components/richTextInput/RichTextEditor.tsx
@@ -79,12 +79,12 @@ const MenuBar = ({ editor }: MenuBarProps) => {
       onClick: () => editor.chain().focus().toggleBold().run(),
       isActive: editor.isActive("bold"),
     },
-    {
-      icon: <FaItalic />,
-      label: "Italic",
-      onClick: () => editor.chain().focus().toggleItalic().run(),
-      isActive: editor.isActive("italic"),
-    },
+    // {
+    //   icon: <FaItalic />,
+    //   label: "Italic",
+    //   onClick: () => editor.chain().focus().toggleItalic().run(),
+    //   isActive: editor.isActive("italic"),
+    // },
     {
       icon: <FaUnderline />,
       label: "Underline",
@@ -236,7 +236,6 @@ const EditorWrapper = styled.div`
 const ControlGroup = styled.div<{ $isSticky: boolean }>`
   position: sticky;
   z-index: 100;
-  margin-bottom: 2px;
   top: 0px;
   transition: all 0.3s ease;
 `;
